+++ core/src/main/java/arez/annotations/Memoize.java
@@ -52,4 +52,16 @@ public @interface Memoize
    * @return the priority of the ComputedValue observer.
    */
   Priority priority() default Priority.NORMAL;
+
+  /**
+   * Flag controlling whether the underlying observer can observe ComputedValue instances with lower priorities.
+   * The default value of false will result in an invariant failure (in development mode) if a lower priority
+   * dependency is observed by the observer. This is to prevent priority inversion when scheduling a higher
+   * priority observer is dependent upon a lower priority computed value. If the value is true then the no
+   * invariant failure is triggered and the component relies on the component author to handle possible priority
+   * inversion.
+   *
+   * @return false if observing lower priority dependencies should result in invariant failure in development mode.
+   */
+  boolean observeLowerPriorityDependencies() default false;
 }