+++ core/src/test/java/arez/component/MemoizeCacheTest.java
@@ -44,12 +44,13 @@ public class MemoizeCacheTest
     final String name = ValueUtil.randomString();
     final Priority priority = Priority.NORMAL;
     final MemoizeCache<String> cache =
-      new MemoizeCache<>( null, null, name, function, 2, priority );
+      new MemoizeCache<>( null, null, name, function, 2, priority, true );
 
     assertEquals( cache.isDisposed(), false );
     assertEquals( cache.getNextIndex(), 0 );
 
     assertEquals( cache.getPriority(), priority );
+    assertEquals( cache.canObserveLowerPriorityDependencies(), true );
 
     final Observer observer1 = context.autorun( () -> {
       observeADependency();
@@ -119,7 +120,7 @@ public class MemoizeCacheTest
     };
     final String name = ValueUtil.randomString();
     final MemoizeCache<String> cache =
-      new MemoizeCache<>( null, null, name, function, 2, Priority.NORMAL );
+      new MemoizeCache<>( null, null, name, function, 2, Priority.NORMAL, false );
 
     final ComputedValue<String> computedValue1 = cache.getComputedValue( "a", "b" );
     final ComputedValue<String> computedValue1b = cache.getComputedValue( "a", "b" );
@@ -153,7 +154,7 @@ public class MemoizeCacheTest
 
     final String name = ValueUtil.randomString();
     final MemoizeCache<String> cache =
-      new MemoizeCache<>( null, component, name, function, 2, Priority.NORMAL );
+      new MemoizeCache<>( null, component, name, function, 2, Priority.NORMAL, false );
 
     final ComputedValue<String> computedValue1 = cache.getComputedValue( "a", "b" );
 
@@ -180,7 +181,7 @@ public class MemoizeCacheTest
     final ArezContext context = Arez.context();
     final String name = ValueUtil.randomString();
     final MemoizeCache<String> cache =
-      new MemoizeCache<>( null, null, name, function, 2, Priority.NORMAL );
+      new MemoizeCache<>( null, null, name, function, 2, Priority.NORMAL, false );
 
     assertEquals( cache.isDisposed(), false );
     assertEquals( cache.getNextIndex(), 0 );
@@ -222,7 +223,7 @@ public class MemoizeCacheTest
     final ArezContext context = Arez.context();
     final String name = ValueUtil.randomString();
     final MemoizeCache<String> cache =
-      new MemoizeCache<>( null, null, name, function, 2, Priority.NORMAL );
+      new MemoizeCache<>( null, null, name, function, 2, Priority.NORMAL, false );
 
     assertEquals( cache.getNextIndex(), 0 );
 
@@ -244,7 +245,8 @@ public class MemoizeCacheTest
                           ValueUtil.randomString(),
                           args -> args[ 0 ] + "." + args[ 1 ],
                           2,
-                          Priority.NORMAL );
+                          Priority.NORMAL,
+                          false );
 
     final IllegalStateException exception =
       expectThrows( IllegalStateException.class, () -> cache.disposeComputedValue( "a" ) );
@@ -262,7 +264,8 @@ public class MemoizeCacheTest
                           ValueUtil.randomString(),
                           args -> args[ 0 ] + "." + args[ 1 ],
                           2,
-                          Priority.NORMAL );
+                          Priority.NORMAL,
+                          false );
 
     final IllegalStateException exception =
       expectThrows( IllegalStateException.class, () -> cache.get( "a" ) );
@@ -275,7 +278,7 @@ public class MemoizeCacheTest
   public void get_invokedOnDisposed()
   {
     final MemoizeCache<String> cache =
-      new MemoizeCache<>( null, null, "X", args -> args[ 0 ] + "." + args[ 1 ], 2, Priority.NORMAL );
+      new MemoizeCache<>( null, null, "X", args -> args[ 0 ] + "." + args[ 1 ], 2, Priority.NORMAL, false );
 
     cache.dispose();
 
@@ -292,7 +295,7 @@ public class MemoizeCacheTest
 
     final IllegalStateException exception =
       expectThrows( IllegalStateException.class,
-                    () -> new MemoizeCache<>( null, null, "X", args -> args[ 0 ], 1, Priority.NORMAL ) );
+                    () -> new MemoizeCache<>( null, null, "X", args -> args[ 0 ], 1, Priority.NORMAL, false ) );
 
     assertEquals( exception.getMessage(),
                   "Arez-0159: MemoizeCache passed a name 'X' but Arez.areNamesEnabled() is false" );
@@ -303,7 +306,7 @@ public class MemoizeCacheTest
   {
     final IllegalStateException exception =
       expectThrows( IllegalStateException.class,
-                    () -> new MemoizeCache<>( null, null, "X", args -> args[ 0 ], 0, Priority.NORMAL ) );
+                    () -> new MemoizeCache<>( null, null, "X", args -> args[ 0 ], 0, Priority.NORMAL, false ) );
 
     assertEquals( exception.getMessage(),
                   "Arez-0160: MemoizeCache constructed with invalid argCount: 0. Expected positive value." );
@@ -316,7 +319,13 @@ public class MemoizeCacheTest
     ArezTestUtil.resetState();
     final IllegalStateException exception =
       expectThrows( IllegalStateException.class,
-                    () -> new MemoizeCache<>( Arez.context(), null, null, args -> args[ 0 ], 1, Priority.NORMAL ) );
+                    () -> new MemoizeCache<>( Arez.context(),
+                                              null,
+                                              null,
+                                              args -> args[ 0 ],
+                                              1,
+                                              Priority.NORMAL,
+                                              false ) );
 
     assertEquals( exception.getMessage(),
                   "Arez-174: MemoizeCache passed a context but Arez.areZonesEnabled() is false" );